server:addedDetails
